Don't document --user installations
Remove the documentation of a --user installation and instead provide
two more options using either cp or a virtualenv. --user is potentially
dangerous as it may install dependencies in a version that is not
compatible with Python tools provided by the Linux package manager,
which are always included on the pythonpath.